Major Rework
Updated and added command line switches for removing apps, disabling
services, and experimenting. Segmented and added switches to interface
and windows update changes.

Added new scripts to disable cortana and start menu annoyances. Also
added a reset script for the start menu to remove dead links from
removed apps.

+++ scripts/remove-default-apps.ps1
@@ -1,38 +1,46 @@
+#   Name: remove-default-apps.ps1
+#
+#   Params:
+#      -System - Include apps from the 'system' group
+#      -Experimental - Include apps from the 'experimental' group
+#      -DryRun - Prints the apps to remove/exclude but does not remove them
+#      -Exclude <array of strings> - Excludes any packages whose name includes an exclude string 
+#
 #   Description:
-# This script removes unwanted Apps that come with Windows. If you  do not want
-# to remove certain Apps comment out the corresponding lines below.
+#      This script removes unwanted Apps that come with Windows. 
+#
+
+Param(
+  [switch]$System,
+  [switch]$Experimental,
+  [switch]$DryRun,
+  [array]$Exclude = $null
+)
 
 Import-Module -DisableNameChecking $PSScriptRoot\..\lib\take-own.psm1
 
 echo "Elevating priviledges for this process"
 do {} until (Elevate-Privileges SeTakeOwnershipPrivilege)
 
-echo "Uninstalling default apps"
 $apps = @(
-    # default Windows 10 apps
+    # Default Windows 10 apps
     "Microsoft.3DBuilder"
     "Microsoft.Appconnector"
     "Microsoft.BingFinance"
     "Microsoft.BingNews"
     "Microsoft.BingSports"
     "Microsoft.BingWeather"
-    #"Microsoft.FreshPaint"
     "Microsoft.Getstarted"
     "Microsoft.MicrosoftOfficeHub"
     "Microsoft.MicrosoftSolitaireCollection"
-    #"Microsoft.MicrosoftStickyNotes"
     "Microsoft.Office.OneNote"
-    #"Microsoft.OneConnect"
     "Microsoft.People"
     "Microsoft.SkypeApp"
-    #"Microsoft.Windows.Photos"
     "Microsoft.WindowsAlarms"
-    #"Microsoft.WindowsCalculator"
     "Microsoft.WindowsCamera"
     "Microsoft.WindowsMaps"
     "Microsoft.WindowsPhone"
     "Microsoft.WindowsSoundRecorder"
-    #"Microsoft.WindowsStore"
     "Microsoft.XboxApp"
     "Microsoft.ZuneMusic"
     "Microsoft.ZuneVideo"
@@ -45,14 +53,13 @@ $apps = @(
     "Microsoft.Messaging"
     "Microsoft.Office.Sway"
 
-
-    #Redstone apps
+    # Redstone apps
     "Microsoft.BingFoodAndDrink"
     "Microsoft.BingTravel"
     "Microsoft.BingHealthAndFitness"
     "Microsoft.WindowsReadingList"
 
-    # non-Microsoft
+    # Non-Microsoft
     "9E2F88E3.Twitter"
     "PandoraMediaInc.29680B314EFC2"
     "Flipboard.Flipboard"
@@ -68,24 +75,67 @@ $apps = @(
     "D52A8D61.FarmVille2CountryEscape"
     "TuneIn.TuneInRadio"
     "GAMELOFTSA.Asphalt8Airborne"
-    #"TheNewYorkTimes.NYTCrossword"
-
-    # apps which cannot be removed using Remove-AppxPackage
-    #"Microsoft.BioEnrollment"
-    #"Microsoft.MicrosoftEdge"
-    #"Microsoft.Windows.Cortana"
-    #"Microsoft.WindowsFeedback"
-    #"Microsoft.XboxGameCallableUI"
-    #"Microsoft.XboxIdentityProvider"
-    #"Windows.ContactSupport"
+    "TheNewYorkTimes.NYTCrossword"
 )
 
+$system_apps = @(
+
+    "Microsoft.MicrosoftStickyNotes"
+    "Microsoft.OneConnect"
+    "Microsoft.Windows.Photos"
+    "Microsoft.WindowsCalculator"
+    "Microsoft.WindowsStore"
+
+)
+
+$experiemental_apps = @(
+
+    "Microsoft.Advertising.Xaml"
+    "Microsoft.PurchaseStoreApp"
+
+)
+
+<#
+$unremovable_apps = @(
+    # apps which cannot be removed using Remove-AppxPackage
+    "Microsoft.BioEnrollment"
+    "Microsoft.MicrosoftEdge"
+    "Microsoft.Windows.Cortana"
+    "Microsoft.WindowsFeedback"
+    "Microsoft.XboxGameCallableUI"
+    "Microsoft.XboxIdentityProvider"
+    "Windows.ContactSupport"
+)
+#>
+
+if($System){$apps += $system_apps}
+if($Experimental){$apps += $experiemental_apps}
+
+if($DryRun) { echo "Listing apps to uninstall" }
+else { echo "Uninstalling default apps" }
+
 foreach ($app in $apps) {
-    echo "Trying to remove $app"
+    
+    $skip = $false
+    foreach($xstring in $Exclude)
+    {
+        if($app.IndexOf($xstring) -ne -1)
+        {
+            echo "Excluding $app"
+            $skip = $true
+        }
+    }
+    
+    if(-not $skip) {
+    
+        echo "Trying to remove $app"
 
-    Get-AppxPackage -Name $app -AllUsers | Remove-AppxPackage
+        if(-not $DryRun) {
+        Get-AppxPackage -Name $app -AllUsers | Remove-AppxPackage
 
-    Get-AppXProvisionedPackage -Online |
-        where DisplayName -EQ $app |
-        Remove-AppxProvisionedPackage -Online
+        Get-AppXProvisionedPackage -Online |
+            where DisplayName -EQ $app |
+            Remove-AppxProvisionedPackage -Online
+        }
+    }
 }
